+++ src/Products/PloneMeeting/migrations/migrate_to_4211.py
@@ -9,6 +9,7 @@ class Migrate_To_4211(Migrator):
     def _updateDataRelatedToToolPloneMeetingSimplification(self):
         """ToolPloneMeeting will be moved to the registry,
            most methods are moved or removed, update stored data."""
+        logger.info('Updating portal_plonemeeting related data...')
         prefixes = ('tool',
                     'context.portal_plonemeeting',
                     'here.portal_plonemeeting',
@@ -30,11 +31,21 @@ class Migrate_To_4211(Migrator):
 
         # POD templates
         self.updatePODTemplatesCode(replacements)
+        logger.info('Done.')
+
+    def _updateItemSearchedSortOn(self):
+        """Make sure every item related searches use sort_on=modified."""
+        logger.info('Updating item searches sort_on...')
+        for cfg in self.tool.objectValues('MeetingConfig'):
+            for collection in cfg.searches.searches_items.objectValues():
+                collection.sort_on = u'modified'
+        logger.info('Done.')
 
     def run(self, extra_omitted=[], from_migration_to_4200=False):
 
         logger.info('Migrating to PloneMeeting 4211...')
         self._updateDataRelatedToToolPloneMeetingSimplification()
+        self._updateItemSearchedSortOn()
         # add text criterion on item title only
         self.updateFacetedFilters(xml_filename='upgrade_step_4211_add_item_widgets.xml')
         logger.info('Migrating to PloneMeeting 4211... Done.')
@@ -44,7 +55,8 @@ def migrate(context):
     '''This migration function will:
 
        1) Update code regarding removal of methods that were available on portal_plonemeeting;
-       2) Add c32 faceted criterion (search on item title only).
+       2) Update every item related searches to use sort_on=modified;
+       3) Add c32 faceted criterion (search on item title only).
     '''
     migrator = Migrate_To_4211(context)
     migrator.run()